Refactor: 모달을 컴포넌트로 생성
- 모달을 컴포넌트에 생성해두고, 불러와서 사용하도록 수정했습니다.
- 모달이 제목을 받도록 하고, 모달 파일 내에 스타일을 포함시켰습니다.
