+++ client/src/components/Modal.tsx
@@ -1,0 +1,63 @@
+import React, { useState } from 'react';
+import styled from 'styled-components';
+
+export const ModalContainer = styled.div`
+  width: 100%;
+  top: -15vh;
+  display: block;
+`;
+
+export const ModalBackground = styled.div`
+  width: 100vw;
+  height: 100vh;
+  top: 0;
+  left: 0;
+  position: absolute;
+  background-color: rgba(0, 0, 0, 0.7);
+  z-index: 2;
+`;
+
+export const ModalWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  padding: 2rem 2rem 1.5rem 2rem;
+  width: 50vw;
+  height: 40vh;
+  top: 30vh;
+  left: 25vw;
+  position: absolute;
+  background-color: ${({ theme }) => theme.colors.black};
+  border-radius: 1rem;
+  z-index: 5;
+`;
+
+export const Title = styled.h4`
+  font-size: ${({ theme }) => theme.fontSizes.xl};
+`;
+
+interface ModalProps {
+  title?: string;
+}
+
+const Modal: React.FC<ModalProps> = ({ children, title }) => {
+  const [modal, setModal] = useState(true);
+  const onClickModalBackground = () => setModal(false);
+  return (
+    <>
+      {modal ? (
+        <ModalContainer>
+          <ModalBackground onClick={onClickModalBackground}>
+            <ModalWrapper onClick={(e) => e.stopPropagation()}>
+              {title ? <Title>{title}</Title> : null}
+              {children}
+            </ModalWrapper>
+          </ModalBackground>
+        </ModalContainer>
+      ) : null}
+    </>
+  );
+};
+
+export default Modal;
