+++ client/src/pages/Login/Login.tsx
@@ -1,24 +1,11 @@
-/* eslint-disable react-hooks/exhaustive-deps */
-import React, { useState } from 'react';
+import React from 'react';
 import useInput from '../../hooks/useInput';
-import {
-  Container,
-  Title,
-  Form,
-  Input,
-  Button,
-  GithubLoginButton,
-  ModalContainer,
-  ModalBackground,
-  ModalWrapper,
-} from './style';
+import { Container, Form, Input, Button, GithubLoginButton } from './style';
+import Modal from '../../components/Modal';
 
 const Login: React.FC = () => {
   const [email, onChangeEmail] = useInput('');
   const [password, onChangePassword] = useInput('');
-  const [modal, setModal] = useState(true);
-
-  const onClickModalBackground = () => setModal(false);
 
   const onClickGithubLogin = () => {
     // Github Login request
@@ -33,41 +20,23 @@ const Login: React.FC = () => {
   };
 
   return (
-    <>
-      {modal ? (
-        <ModalContainer>
-          <ModalBackground onClick={onClickModalBackground}>
-            <ModalWrapper onClick={(e) => e.stopPropagation()}>
-              <Container>
-                <Title>로그인</Title>
-                <Form onSubmit={onSubmitForm}>
-                  <Input
-                    type="text"
-                    placeholder="Email"
-                    id="email"
-                    value={email}
-                    onChange={onChangeEmail}
-                    maxLength={50}
-                  />
-                  <Input
-                    type="password"
-                    placeholder="Password"
-                    id="password"
-                    value={password}
-                    maxLength={15}
-                    onChange={onChangePassword}
-                  />
-                  <Button>로그인</Button>
-                </Form>
-                <GithubLoginButton onClick={onClickGithubLogin}>Github 로그인</GithubLoginButton>
-              </Container>
-            </ModalWrapper>
-          </ModalBackground>
-        </ModalContainer>
-      ) : (
-        <div>로그아웃 모달 제거</div>
-      )}
-    </>
+    <Modal title="로그인">
+      <Container>
+        <Form onSubmit={onSubmitForm}>
+          <Input type="text" placeholder="Email" id="email" value={email} onChange={onChangeEmail} maxLength={50} />
+          <Input
+            type="password"
+            placeholder="Password"
+            id="password"
+            value={password}
+            maxLength={15}
+            onChange={onChangePassword}
+          />
+          <Button>로그인</Button>
+        </Form>
+        <GithubLoginButton onClick={onClickGithubLogin}>Github 로그인</GithubLoginButton>
+      </Container>
+    </Modal>
   );
 };
