+++ client/src/pages/Login/style.ts
@@ -3,12 +3,6 @@ import styled from 'styled-components';
 export const Container = styled.div`
   display: flex;
   flex-direction: column;
-  justify-content: center;
-  align-items: center;
-`;
-
-export const Title = styled.h4`
-  font-size: ${({ theme }) => theme.fontSizes.xl};
 `;
 
 export const Form = styled.form`
@@ -41,35 +35,4 @@ export const GithubLoginButton = styled.button`
   width: 16rem;
   border: 1px solid rgba(0, 0, 0, 0.8);
   border-radius: 1rem;
-`;
-
-export const ModalContainer = styled.div`
-  width: 100%;
-  top: -15vh;
-  display: block;
-`;
-
-export const ModalBackground = styled.div`
-  width: 100vw;
-  height: 100vh;
-  top: 0;
-  left: 0;
-  position: absolute;
-  background-color: rgba(0, 0, 0, 0.7);
-  z-index: 2;
-`;
-
-export const ModalWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  padding: 2rem 2rem 1.5rem 2rem;
-  width: 50vw;
-  height: 40vh;
-  top: 30vh;
-  left: 25vw;
-  position: absolute;
-  background-color: ${({ theme }) => theme.colors.black};
-  border-radius: 1rem;
-  z-index: 5;
 `;
